Replace size_t with uint64_t for detections being decoded inside Python wrapper

+++ src/simplex.pybind.h
@@ -83,7 +83,7 @@ void add_simplex_module(py::module& root) {
       .def(
           "decode_from_detection_events",
           [](SimplexDecoder& self, const py::array_t<bool>& syndrome) {
-            std::vector<size_t> detections;
+            std::vector<uint64_t> detections;
             auto syndrome_unchecked = syndrome.unchecked<1>();
             for (size_t i = 0; i < syndrome_unchecked.size(); i++) {
               if (syndrome_unchecked(i)) {
@@ -127,7 +127,7 @@ void add_simplex_module(py::module& root) {
 
             // Process and decode each shot.
             for (size_t i = 0; i < num_shots; ++i) {
-              std::vector<size_t> detections;
+              std::vector<uint64_t> detections;
               for (size_t j = 0; j < num_detectors; ++j) {
                 if (syndromes_unchecked(i, j)) {
                   detections.push_back(j);
